4.async-python-sprint-4/requirements.txt
aiopg==1.3.5
alembic==1.8.1
anyio==3.6.1
async-timeout==4.0.2
asyncpg==0.26.0
bcrypt==4.0.1
click==8.1.3
Cython==0.29.32
databases==0.6.1
ecdsa==0.18.0
fastapi==0.79.0
greenlet==1.1.3.post0
h11==0.14.0
idna==3.4
Mako==1.2.3
MarkupSafe==2.1.1
orjson==3.8.0
passlib==1.7.4
psycopg2-binary==2.9.4
pyasn1==0.4.8
pydantic==1.9.0
python-dotenv==0.21.0
python-jose==3.3.0
rsa==4.9
setuptools==65.4.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.38.3
starlette==0.19.1
typing_extensions==4.4.0
uvicorn==0.18.2
wheel==0.37.1
